requirements.txt
Flask==2.3.3
google-generativeai==0.3.2
pandas==2.1.4
python-dotenv==1.0.0
